Refactor Navbar component and create ToggleLanguage component
