--- src/components/Navbar/styles.module.scss
@@ -1,7 +1,0 @@
-@import '../../scss/variables/colors';
-
-.container {
-  display: flex;
-  justify-content: flex-end;
-  padding: 10px;
-}
