+++ src/components/ToggleLanguage/styles.module.scss
@@ -1,0 +1,7 @@
+@import '../../scss/variables/colors';
+
+.container {
+  display: flex;
+  justify-content: flex-end;
+  padding: 10px;
+}
